✨ feat: 태블릿일때 반응형 모달 제작 태블릿환경에서 보여지는 모달을 제작했습니다. Resolve: #176
